refactor(analysis): split generation pipeline into modules
Why: analysis-generation.js had become monolithic and expensive to evolve safely.\n\nWhat: extracted focused modules for analysis, role-tag inference, core diagrams, role diagrams, and shared utilities; added dispatcher and graph-utils coverage; and centralized empty-state/seed-merging helpers used by role diagrams.\n\nImpact/Risk: behavior is intended to stay equivalent with lower coupling and smaller edit surfaces; harness memory-gate remains externally blocked by stale FORJAMIE policy state.
