+++ src/core/analysis-generation-analyze-files.js
@@ -1,0 +1,36 @@
+const fs = require('fs');
+const path = require('path');
+const { glob } = require('glob');
+const chalk = require('chalk');
+
+async function resolveCandidateFiles(rootPath, patterns, exclude, explicitFiles) {
+  if (Array.isArray(explicitFiles) && explicitFiles.length > 0) {
+    return [...new Set(explicitFiles
+      .map((filePath) => {
+        const absolute = path.isAbsolute(filePath) ? filePath : path.resolve(rootPath, filePath);
+        const relativeToRoot = path.relative(rootPath, absolute);
+        if (relativeToRoot.startsWith('..') || path.isAbsolute(relativeToRoot)) {
+          return null;
+        }
+        return absolute;
+      })
+      .filter((filePath) => filePath && fs.existsSync(filePath) && fs.statSync(filePath).isFile())
+    )];
+  }
+
+  const files = [];
+  for (const pattern of patterns) {
+    if (!pattern || pattern.trim() === '') continue;
+    try {
+      const matches = await glob(pattern.trim(), { cwd: rootPath, absolute: true, ignore: exclude });
+      files.push(...matches);
+    } catch (_error) {
+      console.warn(chalk.yellow(`⚠️  Invalid pattern: ${pattern}`));
+    }
+  }
+  return [...new Set(files)];
+}
+
+module.exports = {
+  resolveCandidateFiles,
+};
