+++ src/core/analysis-generation-analyze-options.js
@@ -1,0 +1,35 @@
+function parseMaxFiles(options) {
+  let maxFiles = parseInt(options.maxFiles, 10);
+  if (isNaN(maxFiles) || maxFiles < 1 || maxFiles > 10000) {
+    maxFiles = 100;
+  }
+  return Math.min(Math.max(maxFiles, 1), 10000);
+}
+
+function parsePatterns(options) {
+  let patterns = ['**/*.ts', '**/*.tsx', '**/*.js', '**/*.jsx', '**/*.py', '**/*.go', '**/*.rs'];
+  if (options.patterns) {
+    if (typeof options.patterns !== 'string') {
+      throw new TypeError('patterns must be a string');
+    }
+    patterns = options.patterns.split(',');
+  }
+  return patterns;
+}
+
+function parseExclude(options) {
+  let exclude = ['node_modules/**', '.git/**', 'dist/**', 'build/**', '*.test.*', '*.spec.*'];
+  if (options.exclude) {
+    if (typeof options.exclude !== 'string') {
+      throw new TypeError('exclude must be a string');
+    }
+    exclude = options.exclude.split(',');
+  }
+  return exclude;
+}
+
+module.exports = {
+  parseMaxFiles,
+  parsePatterns,
+  parseExclude,
+};
