+++ src/core/analysis-generation-diagrams-role-ai.js
@@ -1,0 +1,11 @@
+const { generateAgent } = require('./analysis-generation-diagrams-role-ai-agent');
+const {
+  generateC4Context,
+  generateRag,
+} = require('./analysis-generation-diagrams-role-ai-context');
+
+module.exports = {
+  generateAgent,
+  generateC4Context,
+  generateRag,
+};
