+++ src/core/analysis-generation-diagrams-role-data.js
@@ -1,0 +1,149 @@
+const {
+  escapeMermaid,
+  componentsByRole,
+  buildRoleDiagramContext,
+  appendDependencyEdges,
+  appendClassAssignment,
+  inferDbIntent,
+} = require('./analysis-generation-utils');
+const { flowNote, noteNode } = require('./analysis-generation-diagrams-empty');
+
+function generateDatabase(data) {
+  if (!data || !Array.isArray(data.components)) {
+    return flowNote('No data available');
+  }
+
+  const lines = ['flowchart TD'];
+  const seeds = componentsByRole(data.components, 'database');
+  if (seeds.length === 0) {
+    lines.push(noteNode('No database-focused components found'));
+    return lines.join('\n');
+  }
+
+  const { byName, safeNames } = buildRoleDiagramContext(data, seeds, 2, 28);
+
+  lines.push('  UserRequest["User request"]');
+  lines.push('  Decision{Record exists?}');
+
+  const addedEdges = new Set();
+  const dbNodeIds = [];
+
+  for (const seed of seeds) {
+    const safe = safeNames.get(seed);
+    if (!safe) continue;
+
+    dbNodeIds.push(safe);
+    lines.push(`  ${safe}["${escapeMermaid(seed.originalName)}"]`);
+    lines.push(`  UserRequest --> ${safe}`);
+
+    const intent = inferDbIntent(seed);
+    if (intent.hasLookup) {
+      const lookup = `${safe}_lookup`;
+      const create = `${safe}_create`;
+      const update = `${safe}_update`;
+      lines.push(`  ${safe} --> ${lookup}["lookup query"]`);
+      lines.push(`  ${lookup} --> Decision`);
+      lines.push(`  Decision -->|found| ${update}["update or modify"]`);
+      lines.push(`  Decision -->|not found| ${create}["insert/create"]`);
+      lines.push(`  ${update} --> ${safe}_result["result"]`);
+      lines.push(`  ${create} --> ${safe}_result["result"]`);
+    } else if (intent.hasWrite) {
+      const write = `${safe}_write`;
+      lines.push(`  ${safe} --> ${write}["write/update"]`);
+      lines.push(`  ${write} --> ${safe}_result["result"]`);
+    } else {
+      const result = `${safe}_result`;
+      lines.push(`  ${safe} --> ${result}["result"]`);
+    }
+  }
+
+  appendDependencyEdges(lines, seeds, byName, safeNames, addedEdges);
+
+  lines.push('  classDef dbNode fill:#0ea5e9,color:#fff');
+  lines.push('  classDef decisionNode fill:#0284c7,color:#fff');
+  appendClassAssignment(lines, dbNodeIds, 'dbNode');
+  lines.push('  class Decision decisionNode');
+  return lines.join('\n');
+}
+
+function generateUserInteractions(data) {
+  if (!data || !Array.isArray(data.components)) {
+    return flowNote('No data available', 'LR');
+  }
+
+  const lines = ['flowchart LR'];
+  const seeds = componentsByRole(data.components, 'user');
+  if (seeds.length === 0) {
+    lines.push(noteNode('No user-facing components found'));
+    return lines.join('\n');
+  }
+
+  const { connected, byName, safeNames } = buildRoleDiagramContext(data, seeds, 1, 30);
+  const edges = new Set();
+  const userNodeIds = [];
+
+  lines.push('  User(("User"))');
+  for (const seed of seeds) {
+    const safe = safeNames.get(seed);
+    if (!safe) continue;
+    userNodeIds.push(safe);
+    lines.push(`  ${safe}["${escapeMermaid(seed.originalName)}"]`);
+    lines.push(`  User --> ${safe}`);
+  }
+
+  appendDependencyEdges(lines, connected, byName, safeNames, edges);
+
+  lines.push('  classDef userNode fill:#16a34a,color:#fff');
+  appendClassAssignment(lines, userNodeIds, 'userNode');
+  return lines.join('\n');
+}
+
+function generateEvents(data) {
+  if (!data || !Array.isArray(data.components)) {
+    return flowNote('No data available');
+  }
+
+  const lines = ['flowchart TD'];
+  const seeds = componentsByRole(data.components, 'events');
+  if (seeds.length === 0) {
+    lines.push(noteNode('No event/channels components found'));
+    return lines.join('\n');
+  }
+
+  const { connected, byName, safeNames } = buildRoleDiagramContext(data, seeds, 2, 30);
+  const edges = new Set();
+  const eventNodeIds = [];
+
+  lines.push('  subgraph Channels["Event channels / queues"]');
+  for (const component of connected) {
+    const safe = safeNames.get(component);
+    if (!safe) continue;
+    const isEventSource = seeds.includes(component);
+    if (isEventSource) {
+      eventNodeIds.push(safe);
+      lines.push(`    ${safe}{{"${escapeMermaid(component.originalName)}"}}`);
+    } else {
+      lines.push(`    ${safe}["${escapeMermaid(component.originalName)}"]`);
+    }
+  }
+  lines.push('  end');
+
+  appendDependencyEdges(
+    lines,
+    connected,
+    byName,
+    safeNames,
+    edges,
+    (component) => (seeds.includes(component) ? 'emit' : 'consume')
+  );
+
+  lines.push('  classDef eventNode fill:#db2777,color:#fff');
+  appendClassAssignment(lines, eventNodeIds, 'eventNode');
+  return lines.join('\n');
+}
+
+module.exports = {
+  generateDatabase,
+  generateUserInteractions,
+  generateEvents,
+};
