+++ src/core/analysis-generation-diagrams-role-security.js
@@ -1,0 +1,103 @@
+const {
+  escapeMermaid,
+  sanitize,
+  componentsByRole,
+  mergeRoleComponents,
+  buildRoleDiagramContext,
+  appendDependencyEdges,
+  appendClassAssignment,
+} = require('./analysis-generation-utils');
+const { collectExternalImports } = require('./analysis-generation-role-tags');
+const { flowNote, noteNode } = require('./analysis-generation-diagrams-empty');
+
+function generateAuth(data) {
+  if (!data || !Array.isArray(data.components)) {
+    return flowNote('No data available');
+  }
+
+  const lines = ['flowchart TD'];
+  const seeds = componentsByRole(data.components, 'auth');
+  if (seeds.length === 0) {
+    lines.push(noteNode('No authentication components found'));
+    return lines.join('\n');
+  }
+
+  const { connected, byName, safeNames } = buildRoleDiagramContext(data, seeds, 2, 24);
+  const edges = new Set();
+  const authNodeIds = [];
+
+  lines.push('  Request["Authentication request"]');
+  lines.push('  Boundary{"Auth Boundary"}');
+  lines.push('  Request --> Boundary');
+
+  for (const seed of seeds) {
+    const safe = safeNames.get(seed);
+    if (!safe) continue;
+    authNodeIds.push(safe);
+    lines.push(`  ${safe}["${escapeMermaid(seed.originalName)}"]`);
+    const key = `Boundary->${safe}`;
+    if (!edges.has(key)) {
+      edges.add(key);
+      lines.push(`  Boundary --> ${safe}`);
+    }
+  }
+
+  appendDependencyEdges(lines, connected, byName, safeNames, edges);
+
+  const providerSet = new Set();
+  for (const seed of seeds) {
+    for (const pkg of collectExternalImports(seed.imports || [])) {
+      providerSet.add(pkg);
+    }
+  }
+  for (const provider of providerSet) {
+    const providerNode = sanitize(provider);
+    lines.push(`  ${providerNode}[("${escapeMermaid(provider)}")]`);
+  }
+
+  lines.push('  classDef authNode fill:#7c3aed,color:#fff');
+  appendClassAssignment(lines, authNodeIds, 'authNode');
+  return lines.join('\n');
+}
+
+function generateSecurity(data) {
+  if (!data || !Array.isArray(data.components)) {
+    return flowNote('No data available');
+  }
+
+  const lines = ['flowchart TD'];
+  const seeds = mergeRoleComponents(data.components, ['security', 'auth', 'integrations']);
+
+  if (seeds.length === 0) {
+    lines.push(noteNode('No security-focused components found'));
+    return lines.join('\n');
+  }
+
+  const { connected, byName, safeNames } = buildRoleDiagramContext(data, seeds, 2, 40);
+  const edges = new Set();
+  const securityNodeIds = [];
+
+  lines.push('  Untrusted["Untrusted input"]');
+  for (const seed of seeds) {
+    const safe = safeNames.get(seed);
+    if (!safe) continue;
+    securityNodeIds.push(safe);
+    lines.push(`  ${safe}["${escapeMermaid(seed.originalName)}"]`);
+    const key = `Untrusted->${safe}`;
+    if (!edges.has(key)) {
+      edges.add(key);
+      lines.push(`  Untrusted --> ${safe}`);
+    }
+  }
+
+  appendDependencyEdges(lines, connected, byName, safeNames, edges);
+
+  lines.push('  classDef securityNode fill:#dc2626,color:#fff');
+  appendClassAssignment(lines, securityNodeIds, 'securityNode');
+  return lines.join('\n');
+}
+
+module.exports = {
+  generateAuth,
+  generateSecurity,
+};
