+++ src/core/analysis-generation-role-tags-imports.js
@@ -1,0 +1,26 @@
+const {
+  getImportPath,
+  getExternalPackageName,
+} = require('./analysis-generation-utils');
+
+function collectExternalImports(importEntries) {
+  const packages = new Set();
+  if (!Array.isArray(importEntries)) return [];
+
+  for (const entry of importEntries) {
+    const importPath = getImportPath(entry);
+    if (!importPath || importPath.startsWith('.')) {
+      continue;
+    }
+    const externalPackage = getExternalPackageName(importPath);
+    if (externalPackage) {
+      packages.add(externalPackage);
+    }
+  }
+
+  return [...packages];
+}
+
+module.exports = {
+  collectExternalImports,
+};
