+++ src/core/analysis-generation-role-tags-match.js
@@ -1,0 +1,9 @@
+function textHasToken(text, token) {
+  const escaped = token.replace(/[.*+?^${}()|[\]\\]/g, '\\$&');
+  const re = new RegExp(`(^|[\\/._-])${escaped}([\\/._-]|$)`, 'i');
+  return re.test(text);
+}
+
+module.exports = {
+  textHasToken,
+};
