+++ src/core/analysis-generation-role-tags.js
@@ -1,0 +1,7 @@
+const { collectExternalImports } = require('./analysis-generation-role-tags-imports');
+const { inferRoleTags } = require('./analysis-generation-role-tags-infer');
+
+module.exports = {
+  collectExternalImports,
+  inferRoleTags,
+};
