+++ src/core/analysis-generation-utils-core.js
@@ -1,0 +1,137 @@
+const crypto = require('crypto');
+const path = require('path');
+
+function detectLanguage(filePath) {
+  if (typeof filePath !== 'string') return 'unknown';
+  const ext = path.extname(filePath).toLowerCase();
+  const map = {
+    '.ts': 'typescript', '.tsx': 'typescript',
+    '.mts': 'typescript', '.cts': 'typescript',
+    '.js': 'javascript', '.jsx': 'javascript',
+    '.mjs': 'javascript', '.cjs': 'javascript',
+    '.py': 'python', '.go': 'go', '.rs': 'rust',
+    '.java': 'java', '.rb': 'ruby', '.php': 'php',
+  };
+  return map[ext] || 'unknown';
+}
+
+function inferType(filePath, content) {
+  const base = path.basename(filePath).toLowerCase();
+  if (base.includes('service')) return 'service';
+  if (base.includes('component') || base.endsWith('.tsx') || base.endsWith('.jsx')) return 'component';
+  if (content.includes('class ') && content.includes('extends')) return 'class';
+  if (content.includes('export default function') || content.includes('export function')) return 'function';
+  if (content.includes('module.exports') || content.includes('export ')) return 'module';
+  return 'file';
+}
+
+function extractImports(content, lang) {
+  const imports = [];
+  if (lang === 'typescript' || lang === 'javascript') {
+    const es6Regex = /import\s+(?:(?:\{[^}]*?\}|\*\s+as\s+\w+|\w+)\s+from\s+)?["']([^"']+)["']/g;
+    const es6 = [...content.matchAll(es6Regex)];
+    es6.forEach(m => imports.push(m[1]));
+
+    const cjs = [...content.matchAll(/require\s*\(\s*["']([^"']+)["']\s*\)/g)];
+    cjs.forEach(m => imports.push(m[1]));
+
+    const dynamic = [...content.matchAll(/import\s*\(\s*["']([^"']+)["']\s*\)/g)];
+    dynamic.forEach(m => imports.push(m[1]));
+  } else if (lang === 'python') {
+    const py = [...content.matchAll(/(?:from|import)\s+([\w.]+)/g)];
+    py.forEach(m => imports.push(m[1]));
+  } else if (lang === 'go') {
+    const go = [...content.matchAll(/import\s+(?:\(\s*)?["']([^"']+)["']/g)];
+    go.forEach(m => imports.push(m[1]));
+  }
+  return imports;
+}
+
+function extractImportsWithPositions(content, lang) {
+  const imports = [];
+  const lines = content.split(/\r?\n/);
+
+  for (let i = 0; i < lines.length; i++) {
+    const line = lines[i];
+    const lineNum = i + 1;
+
+    if (lang === 'typescript' || lang === 'javascript') {
+      const es6 = line.match(/import\s+(?:(?:\{[^}]*\}|\*\s+as\s+\w+|\w+)\s+from\s+)?["']([^"']+)["']/);
+      if (es6) {
+        imports.push({ path: es6[1], line: lineNum });
+        continue;
+      }
+
+      const cjs = line.match(/require\s*\(\s*["']([^"']+)["']\s*\)/);
+      if (cjs) {
+        imports.push({ path: cjs[1], line: lineNum });
+        continue;
+      }
+
+      const dynamic = line.match(/import\s*\(\s*["']([^"']+)["']\s*\)/);
+      if (dynamic) {
+        imports.push({ path: dynamic[1], line: lineNum });
+      }
+    } else if (lang === 'python') {
+      const py = line.match(/(?:from|import)\s+([\w.]+)/);
+      if (py) {
+        imports.push({ path: py[1], line: lineNum });
+      }
+    } else if (lang === 'go') {
+      const go = line.match(/import\s+(?:\(\s*)?["']([^"']+)["']/);
+      if (go) {
+        imports.push({ path: go[1], line: lineNum });
+      }
+    }
+  }
+
+  return imports;
+}
+
+function sanitize(name) {
+  const base = name.replace(/[^a-zA-Z0-9_]/g, '_').replace(/^[0-9]/, '_$&');
+  const hash = crypto.createHash('sha256').update(name).digest('hex').slice(0, 8);
+  return `${base}_${hash}`;
+}
+
+function escapeMermaid(str) {
+  if (!str) return '';
+  return str.replace(/[\\"\[\]()#<>{}|]/g, '\\$&');
+}
+
+function normalizePath(inputPath) {
+  return inputPath.replace(/\\/g, '/');
+}
+
+function toComparablePath(filePath) {
+  return normalizePath(String(filePath || '')).replace(/^\.\//, '');
+}
+
+function getImportPath(importInfo) {
+  if (typeof importInfo === 'string') return importInfo;
+  if (importInfo && typeof importInfo.path === 'string') return importInfo.path;
+  return null;
+}
+
+function getExternalPackageName(importPath) {
+  if (typeof importPath !== 'string') return null;
+  if (!importPath) return null;
+  if (importPath.startsWith('@')) {
+    const [scope, pkg] = importPath.split('/');
+    return scope && pkg ? `${scope}/${pkg}` : scope || null;
+  }
+  return importPath.split('/')[0] || null;
+}
+
+module.exports = {
+  detectLanguage,
+  inferType,
+  extractImports,
+  extractImportsWithPositions,
+  sanitize,
+  escapeMermaid,
+  normalizePath,
+  toComparablePath,
+  getImportPath,
+  getExternalPackageName,
+};
