+++ src/core/analysis-generation-utils-resolution.js
@@ -1,0 +1,57 @@
+const path = require('path');
+const { normalizePath, toComparablePath } = require('./analysis-generation-utils-core');
+
+const IMPORT_RESOLUTION_SUFFIXES = [
+  '',
+  '.ts',
+  '.tsx',
+  '.js',
+  '.jsx',
+  '.mjs',
+  '.mts',
+  '.cts',
+  '/index.ts',
+  '/index.tsx',
+  '/index.js',
+  '/index.jsx',
+  '/index.mjs',
+  '/index.mts',
+  '/index.cts'
+];
+
+function resolveInternalImport(fromFilePath, importPath, rootPath) {
+  if (typeof fromFilePath !== 'string' || typeof importPath !== 'string') {
+    return null;
+  }
+  if (!importPath.startsWith('.')) {
+    return null;
+  }
+
+  const fromDir = path.dirname(fromFilePath);
+
+  if (rootPath) {
+    const absoluteTarget = path.resolve(rootPath, fromDir, importPath);
+    const relativeToRoot = toComparablePath(path.relative(rootPath, absoluteTarget));
+    if (relativeToRoot.startsWith('..') || path.isAbsolute(relativeToRoot)) {
+      return null;
+    }
+    return relativeToRoot;
+  }
+
+  const posixFromDir = normalizePath(fromDir);
+  const posixImport = normalizePath(importPath);
+  return toComparablePath(path.posix.normalize(path.posix.join(posixFromDir, posixImport)));
+}
+
+function findComponentByResolvedPath(components, resolvedPath) {
+  const comparablePath = toComparablePath(resolvedPath);
+  const candidates = new Set(
+    IMPORT_RESOLUTION_SUFFIXES.map((suffix) => toComparablePath(comparablePath + suffix))
+  );
+  return components.find((component) => candidates.has(toComparablePath(component.filePath)));
+}
+
+module.exports = {
+  resolveInternalImport,
+  findComponentByResolvedPath,
+};
