+++ src/core/analysis-generation-utils.js
@@ -1,0 +1,9 @@
+const coreUtils = require('./analysis-generation-utils-core');
+const resolutionUtils = require('./analysis-generation-utils-resolution');
+const graphUtils = require('./analysis-generation-utils-graph');
+
+module.exports = {
+  ...coreUtils,
+  ...resolutionUtils,
+  ...graphUtils,
+};
